requirements.txt
streamlit
crewai
python-dotenv
langchain-openai
plotly
fuzzywuzzy
python-Levenshtein